refactor: move init code to mocha hooks
